Convert remove-slots to TypeScript
